refactored
factored out self.currentPort and self.portRect.
Now these are passed using graphics view insted of elements. Reason being
we don't want to save these attrs.  We want to slim down element and link
modules as much as possible and only keep the data that we need to save.
This will make it much easier when we actualy save the classes.