Text show in first click/scene cleard on reload
Now when an element is placed the text is shown on first click
also one text is added on element add.  When making a new image
the text is added to the scene before the image finishes valid
pos check so we must explictly set the parent of the test image
to None which is now done with the image.remove() method.

When we load a saved file, the scene is cleard of all elements.